feat(main): create a Menu restaurant

+++ app/src/main/java/org/example/models/ReviewDish.java
@@ -1,12 +1,12 @@
 package org.example.models;
 
 public class ReviewDish implements IReview {
-    private Double tasteRating;
-    private Double presentationRating;
+    private Float tasteRating;
+    private Float presentationRating;
     private String comment;
 
 
-    public ReviewDish(String comment, Double tasteRating, Double presentationRating) {
+    public ReviewDish(String comment, Float tasteRating, Float presentationRating) {
         this.comment=comment;
         this.tasteRating = tasteRating;
         this.presentationRating = presentationRating;
@@ -16,19 +16,19 @@ public class ReviewDish implements IReview {
 
     }
 
-    public Double getTasteRating() {
+    public Float getTasteRating() {
         return tasteRating;
     }
 
-    public void setTasteRating(Double tasteRating) {
+    public void setTasteRating(Float tasteRating) {
         this.tasteRating = tasteRating;
     }
 
-    public Double getPresentationRating() {
+    public Float getPresentationRating() {
         return presentationRating;
     }
 
-    public void setPresentationRating(Double presentationRating) {
+    public void setPresentationRating(Float presentationRating) {
         this.presentationRating = presentationRating;
     }
